Support iceberg table collect stats.
